Restructure data generator child classes
