+++ data.sql
@@ -31,24 +31,24 @@ VALUES (1, '2022-04-01 23:56:20.061855', '2022-04-01 23:56:20.061855', 'Hello th
        (7, '2022-04-01 23:52:38.311749', '2022-04-01 23:56:20.061855', 'Hello this is caption', 0, 3, NULL, true),
        (8, '2022-04-02 00:02:45.371366', '2022-04-01 23:56:20.061855', 'Hello this is caption', 1, 4, NULL, true),
        (9, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 5, NULL, true),
-        (10, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-         (11, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-          (12, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-           (13, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-            (14, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-             (15, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-              (16, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-               (17, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (18, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (19, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (20, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (21, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (22, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (23, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (24, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (25, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (26, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
-                (27, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true);
+       (10, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (11, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (12, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (13, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (14, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (15, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (16, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (17, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (18, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (19, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (20, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (21, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (22, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (23, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (24, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (25, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (26, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true),
+       (27, '2022-04-02 00:08:00.991663', '2022-04-01 23:56:20.061855', 'Hello this is caption', 2, 1, NULL, true);
 
 
 SELECT setval('post_id_seq', (SELECT MAX(id) FROM post) + 1);
